Move blacken&flynt test to correct module
